# apps/api-gateway/openapi/openapi.yaml
openapi: 3.0.0
info:
  title: AP Calculus Bot API
  description: API Gateway for AP Calculus AB/BC tutoring bot with verified answers
  version: 0.1.0
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /coach/:
    post:
      tags:
        - coach
      description: Process a student question using VAM (Verified Answer Mode)
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                subject:
                  type: string
                  enum:
                    - calc
                  default: calc
                examVariant:
                  type: string
                  enum:
                    - calc_ab
                    - calc_bc
                  default: calc_ab
                mode:
                  type: string
                  enum:
                    - vam
                    - standard
                  default: vam
                question:
                  type: string
                  minLength: 1
                  maxLength: 2000
                context:
                  type: object
                  properties:
                    topic:
                      type: string
                    subtopic:
                      type: string
                    difficulty:
                      type: string
                    studentLevel:
                      type: string
                      enum:
                        - beginner
                        - intermediate
                        - advanced
                    previousQuestions:
                      type: array
                      items:
                        type: string
                    sessionId:
                      type: string
                  additionalProperties: false
              required:
                - question
              additionalProperties: false
        required: true
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  answer:
                    type: string
                  verified:
                    type: boolean
                  trustScore:
                    type: number
                  confidence:
                    type: number
                  sources:
                    type: array
                    items:
                      type: object
                      properties:
                        type:
                          type: string
                          enum:
                            - canonical
                            - retrieval
                            - generated
                        id:
                          type: string
                        title:
                          type: string
                        snippet:
                          type: string
                        score:
                          type: number
                      required:
                        - type
                        - id
                      additionalProperties: false
                    default: []
                  suggestions:
                    type: array
                    items:
                      type: string
                    default: []
                  metadata:
                    type: object
                    properties:
                      examVariant:
                        type: string
                      topic:
                        type: string
                      subtopic:
                        type: string
                      difficulty:
                        type: string
                      processingTime:
                        type: number
                      retryCount:
                        type: number
                    additionalProperties: false
                required:
                  - answer
                  - verified
                  - trustScore
                  - confidence
                additionalProperties: false
        "400":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      message:
                        type: string
                        description: Human readable error message
                      statusCode:
                        type: number
                        description: HTTP status code associated with the error
                      code:
                        type: string
                        description: Application-specific error code
                      validation:
                        description: Validation errors when applicable
                      details:
                        description: Additional error context
                    required:
                      - message
                      - statusCode
                    additionalProperties: false
                required:
                  - error
                additionalProperties: false
        "500":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      message:
                        type: string
                        description: Human readable error message
                      statusCode:
                        type: number
                        description: HTTP status code associated with the error
                      code:
                        type: string
                        description: Application-specific error code
                      validation:
                        description: Validation errors when applicable
                      details:
                        description: Additional error context
                    required:
                      - message
                      - statusCode
                    additionalProperties: false
                required:
                  - error
                additionalProperties: false
...
  /coach/config:
    get:
      tags:
        - coach
      description: Get coach configuration
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  vam:
                    type: object
                    properties:
                      minTrustThreshold:
                        type: number
                      maxRetries:
                        type: number
                      enableCanonicalFirst:
                        type: boolean
                      enableRetrieval:
                        type: boolean
                      enableVerification:
                        type: boolean
                      enablePostprocessing:
                        type: boolean
                      cacheVerifiedOnly:
                        type: boolean
                      suggestionsCount:
                        type: number
                    required:
                      - minTrustThreshold
                      - maxRetries
                      - enableCanonicalFirst
                      - enableRetrieval
                      - enableVerification
                      - enablePostprocessing
                      - cacheVerifiedOnly
                      - suggestionsCount
                    additionalProperties: false
                  supportedSubjects:
                    type: array
                    items:
                      type: string
                  supportedVariants:
                    type: array
                    items:
                      type: string
                required:
                  - vam
                  - supportedSubjects
                  - supportedVariants
                additionalProperties: false
        "500":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      message:
                        type: string
                        description: Human readable error message
                      statusCode:
                        type: number
                        description: HTTP status code associated with the error
                      code:
                        type: string
                        description: Application-specific error code
                      validation:
                        description: Validation errors when applicable
                      details:
                        description: Additional error context
                    required:
                      - message
                      - statusCode
                    additionalProperties: false
                required:
                  - error
                additionalProperties: false
...
servers:
  - url: http://localhost:3001
    description: Development server